+++ extract_msg/msg_classes/calendar_base.py
@@ -33,7 +33,7 @@ class CalendarBase(MessageBase):
         """
         A list of all attendees, excluding the organizer.
         """
-        return self._getNamedAs('8238', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8238', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def appointmentAuxilaryFlags(self) -> Optional[AppointmentAuxilaryFlag]:
@@ -54,14 +54,14 @@ class CalendarBase(MessageBase):
         """
         The length of the event, in minutes.
         """
-        return self._getNamedAs('8213', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8213', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def appointmentEndWhole(self) -> Optional[datetime.datetime]:
         """
         The end date and time of the event in UTC.
         """
-        return self._getNamedAs('820E', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('820E', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def appointmentNotAllowPropose(self) -> bool:
@@ -69,7 +69,7 @@ class CalendarBase(MessageBase):
         Indicates that attendees are not allowed to propose a new date and/or
         time for the meeting if True.
         """
-        return self._getNamedAs('8259', ps.PSETID_APPOINTMENT, bool, False)
+        return bool(self.getNamedProp('8259', ps.PSETID_APPOINTMENT))
 
     @functools.cached_property
     def appointmentRecur(self) -> Optional[RecurrencePattern]:
@@ -86,14 +86,14 @@ class CalendarBase(MessageBase):
         begins with the sequence number set to 0 and is incremented each time
         the organizer sends out a Meeting Update object.
         """
-        return self._getNamedAs('8201', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8201', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def appointmentStartWhole(self) -> Optional[datetime.datetime]:
         """
         The start date and time of the event in UTC.
         """
-        return self._getNamedAs('820D', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('820D', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def appointmentStateFlags(self) -> Optional[AppointmentStateFlag]:
@@ -107,7 +107,7 @@ class CalendarBase(MessageBase):
         """
         Whether the event is an all-day event or not.
         """
-        return self._getNamedAs('8215', ps.PSETID_APPOINTMENT, bool, False)
+        return bool(self.getNamedProp('8215', ps.PSETID_APPOINTMENT))
 
     @functools.cached_property
     def appointmentTimeZoneDefinitionEndDisplay(self) -> Optional[TimeZoneDefinition]:
@@ -142,7 +142,7 @@ class CalendarBase(MessageBase):
         the specifications. If you have examples, let me know and I can ask you
         to run a verification on it.
         """
-        return self._getNamedAs('825D', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('825D', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def bcc(self) -> Optional[str]:
@@ -156,7 +156,7 @@ class CalendarBase(MessageBase):
         """
         Indicated the name of the contact associated with the birthday event.
         """
-        return self._getNamedAs('BirthdayContactAttributionDisplayName', ps.PSETID_ADDRESS)
+        return self.getNamedProp('BirthdayContactAttributionDisplayName', ps.PSETID_ADDRESS)
 
     @functools.cached_property
     def birthdayContactEntryID(self) -> Optional[EntryID]:
@@ -171,7 +171,7 @@ class CalendarBase(MessageBase):
         Indicates the person ID's GUID of the contact associated with the
         birthday event.
         """
-        return self._getNamedAs('BirthdayContactPersonGuid', ps.PSETID_ADDRESS)
+        return self.getNamedProp('BirthdayContactPersonGuid', ps.PSETID_ADDRESS)
 
     @functools.cached_property
     def busyStatus(self) -> Optional[BusyStatus]:
@@ -193,7 +193,7 @@ class CalendarBase(MessageBase):
         """
         A list of all the sendable attendees, who are also optional attendees.
         """
-        return self._getNamedAs('823C', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('823C', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def cleanGlobalObjectID(self) -> Optional[GlobalObjectID]:
@@ -215,7 +215,7 @@ class CalendarBase(MessageBase):
 
         Honestly, not sure what this is. [MS-OXOCAL]: PidLidClipEnd.
         """
-        return self._getNamedAs('8236', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8236', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def clipStart(self) -> Optional[datetime.datetime]:
@@ -226,14 +226,14 @@ class CalendarBase(MessageBase):
 
         Honestly, not sure what this is. [MS-OXOCAL]: PidLidClipStart.
         """
-        return self._getNamedAs('8235', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8235', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def commonEnd(self) -> Optional[datetime.datetime]:
         """
         The end date and time of an event. MUST be equal to appointmentEndWhole.
         """
-        return self._getNamedAs('8517', ps.PSETID_COMMON)
+        return self.getNamedProp('8517', ps.PSETID_COMMON)
 
     @functools.cached_property
     def commonStart(self) -> Optional[datetime.datetime]:
@@ -241,14 +241,14 @@ class CalendarBase(MessageBase):
         The start date and time of an event. MUST be equal to
         appointmentStartWhole.
         """
-        return self._getNamedAs('8516', ps.PSETID_COMMON)
+        return self.getNamedProp('8516', ps.PSETID_COMMON)
 
     @functools.cached_property
     def endDate(self) -> Optional[datetime.datetime]:
         """
         The end date of the appointment.
         """
-        return self.props.getValue('00610040')
+        return self.getPropertyVal('00610040')
 
     @functools.cached_property
     def globalObjectID(self) -> Optional[GlobalObjectID]:
@@ -270,29 +270,28 @@ class CalendarBase(MessageBase):
         Indicates whether the contact associated with the birthday event is
         writable.
         """
-        return self._getNamedAs('IsBirthdayContactWritable', ps.PSETID_ADDRESS, bool, False)
-
+        return bool(self.getNamedProp('IsBirthdayContactWritable', ps.PSETID_ADDRESS))
     @functools.cached_property
     def isException(self) -> bool:
         """
         Whether the object represents an exception. False indicates that the
         object represents a recurring series or a single-instance object.
         """
-        return self._getNamedAs('000A', ps.PSETID_MEETING, bool, False)
+        return bool(self.getNamedProp('000A', ps.PSETID_MEETING))
 
     @functools.cached_property
     def isRecurring(self) -> bool:
         """
         Whether the object is associated with a recurring series.
         """
-        return self._getNamedAs('0005', ps.PSETID_MEETING, bool, False)
+        return bool(self.getNamedProp('0005', ps.PSETID_MEETING))
 
     @functools.cached_property
     def keywords(self) -> Optional[List[str]]:
         """
         The color to be used when displaying a Calendar object.
         """
-        return self._getNamedAs('Keywords', ps.PS_PUBLIC_STRINGS)
+        return self.getNamedProp('Keywords', ps.PS_PUBLIC_STRINGS)
 
     @functools.cached_property
     def linkedTaskItems(self) -> Optional[List[EntryID]]:
@@ -307,14 +306,14 @@ class CalendarBase(MessageBase):
         """
         Returns the location of the meeting.
         """
-        return self._getNamedAs('8208', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8208', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def meetingDoNotForward(self) -> bool:
         """
         Whether to allow the meeting to be forwarded. True disallows forwarding.
         """
-        return self._getNamedAs('DoNotForward', ps.PS_PUBLIC_STRINGS, bool, False)
+        return bool(self.getNamedProp('DoNotForward', ps.PS_PUBLIC_STRINGS))
 
     @functools.cached_property
     def meetingWorkspaceUrl(self) -> Optional[str]:
@@ -322,28 +321,28 @@ class CalendarBase(MessageBase):
         The URL of the Meeting Workspace, as specified in [MS-MEETS], that is
         associated with a Calendar object.
         """
-        return self._getNamedAs('8209', ps.PSETID_APPOINTMENT)
+        return self.getNamedProp('8209', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def nonSendableBcc(self) -> Optional[str]:
         """
         A list of all unsendable attendees who are also resource objects.
         """
-        return self._getNamedAs('8538', ps.PSETID_COMMON)
+        return self.getNamedProp('8538', ps.PSETID_COMMON)
 
     @functools.cached_property
     def nonSendableCc(self) -> Optional[str]:
         """
         A list of all unsendable attendees who are also optional attendees.
         """
-        return self._getNamedAs('8537', ps.PSETID_COMMON)
+        return self.getNamedProp('8537', ps.PSETID_COMMON)
 
     @functools.cached_property
     def nonSendableTo(self) -> Optional[str]:
         """
         A list of all unsendable attendees who are also required attendees.
         """
-        return self._getNamedAs('8536', ps.PSETID_COMMON)
+        return self.getNamedProp('8536', ps.PSETID_COMMON)
 
     @functools.cached_property
     def nonSendBccTrackStatus(self) -> Optional[List[ResponseStatus]]:
@@ -371,7 +370,7 @@ class CalendarBase(MessageBase):
         """
         Returns the optional attendees of the meeting.
         """
-        return self._getNamedAs('0007', ps.PSETID_MEETING)
+        return self.getNamedProp('0007', ps.PSETID_MEETING)
 
     @property
     def organizer(self) -> Optional[str]:
@@ -387,7 +386,7 @@ class CalendarBase(MessageBase):
         Assists a client or server in finding a Calendar object but is not
         guarenteed to be unique amoung all objects.
         """
-        return self.props.getValue('00620003')
+        return self.getPropertyVal('00620003')
 
     @functools.cached_property
     def ownerCriticalChange(self) -> Optional[datetime.datetime]:
@@ -395,7 +394,7 @@ class CalendarBase(MessageBase):
         The date and time at which a Meeting Request object was sent by the
         organizer, in UTC.
         """
-        return self.namedProperties.get(('001A', ps.PSETID_MEETING))
+        return self.getNamedProp('001A', ps.PSETID_MEETING)
 
     @functools.cached_property
     def recurrencePattern(self) -> Optional[str]:
@@ -403,56 +402,56 @@ class CalendarBase(MessageBase):
         A description of the recurrence specified by the appointmentRecur
         property.
         """
-        return self.namedProperties.get(('8232', ps.PSETID_APPOINTMENT))
+        return self.getNamedProp('8232', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def recurring(self) -> bool:
         """
         Specifies whether the object represents a recurring series.
         """
-        return bool(self.namedProperties.get(('8223', ps.PSETID_APPOINTMENT), False))
+        return bool(self.getNamedProp('8223', ps.PSETID_APPOINTMENT))
 
     @functools.cached_property
     def replyRequested(self) -> bool:
         """
         Whether the organizer requests a reply from attendees.
         """
-        return bool(self.props.getValue('0C17000B'))
+        return bool(self.getPropertyVal('0C17000B'))
 
     @functools.cached_property
     def requiredAttendees(self) -> Optional[str]:
         """
         Returns the required attendees of the meeting.
         """
-        return self.namedProperties.get(('0006', ps.PSETID_MEETING))
+        return self.getNamedProp('0006', ps.PSETID_MEETING)
 
     @functools.cached_property
     def resourceAttendees(self) -> Optional[str]:
         """
         Returns the resource attendees of the meeting.
         """
-        return self.namedProperties.get(('0008', ps.PSETID_MEETING))
+        return self.getNamedProp('0008', ps.PSETID_MEETING)
 
     @functools.cached_property
     def responseRequested(self) -> bool:
         """
         Whether to send Meeting Response objects to the organizer.
         """
-        return bool(self.props.getValue('0063000B'))
+        return bool(self.getPropertyVal('0063000B'))
 
     @functools.cached_property
     def responseStatus(self) -> ResponseStatus:
         """
         The response status of an attendee.
         """
-        return ResponseStatus(self.namedProperties.get(('8218', ps.PSETID_APPOINTMENT), 0))
+        return ResponseStatus(self.getNamedProp('8218', ps.PSETID_APPOINTMENT, 0))
 
     @functools.cached_property
     def startDate(self) -> Optional[datetime.datetime]:
         """
         The start date of the appointment.
         """
-        return self.props.getValue('00600040')
+        return self.getPropertyVal('00600040')
 
     @functools.cached_property
     def timeZoneDescription(self) -> Optional[str]:
@@ -460,7 +459,7 @@ class CalendarBase(MessageBase):
         A human-readable description of the time zone that is represented by the
         data in the timeZoneStruct property.
         """
-        return self.namedProperties.get(('8234', ps.PSETID_APPOINTMENT))
+        return self.getNamedProp('8234', ps.PSETID_APPOINTMENT)
 
     @functools.cached_property
     def timeZoneStruct(self) -> Optional[TimeZoneStruct]:
@@ -482,4 +481,4 @@ class CalendarBase(MessageBase):
         """
         A list of all the sendable attendees, who are also required attendees.
         """
-        return self.namedProperties.get(('823B', ps.PSETID_APPOINTMENT))
+        return self.getNamedProp('823B', ps.PSETID_APPOINTMENT)